% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{lexi}
\alias{lexi}
\title{LEXI: lncRNA expression imputation}
\usage{
lexi(train_pcg, train_lnc, my_pcg, gene_index, num = 100)
}
\arguments{
\item{train_pcg}{training protein coding dataset. a numeric matrix with with row names indicating samples, and column names indicating protein coding gene IDs.}

\item{train_lnc}{training lncRNA expression dataset. a numeric matrix with row names indicating samples, and column names indicating lncRNA IDs}

\item{my_pcg}{test protein coding expression dataset. a numberic matrix with row names indicating samples, and column names indicating protein coding gene IDs.}

\item{gene_index}{either gene name (character) or index (column number) of lncRNA to be imputed.}

\item{num}{number of informative protein coding genes to be used in constructing imputation model. Default is 100 genes.}
}
\value{
a numeric vector containing imputed & standardized expression levels of the lncRNA
}
\description{
Function to impute lncRNA expression profile from protein coding expression dataset
}
\examples{
lexi(train_pcg, train_lnc, my_pcg, gene_index="ENSG00000228630", num=100)
lexi(train_pcg, train_lnc, my_pcg, gene_index=25, num=100)

}
