% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_functions.R
\name{posexp}
\alias{posexp}
\title{Filter lncRNA dataset based on positive expression in a specific percentage of samples}
\usage{
posexp(train_lnc, cutoff = 75, threshold = 1)
}
\arguments{
\item{train_lnc}{a numeric matrix with lncRNA expression data}

\item{cutoff}{percentage of samples in which the lncRNA should be expressed}

\item{threshold}{the numeric threshold that defines whether the lncRNA is expressed or not. Defaults to 1 (for RPKM or FPKM)}
}
\value{
filtered lncRNA expression matrix
}
\description{
Function to return lncRNA expression subset containing lncRNAs that are expressed in a specific percentage of samples
}
\examples{
posexp(train_lnc, cutoff=75, threshold=1)
}
